Restructure iOS folder structure
