+++ Sources/iOS/Classes/CarouselComposite.swift
