+++ Sources/iOS/Classes/CarouselSpotCell.swift
