+++ Sources/iOS/Classes/CarouselSpotHeader.swift
