+++ Sources/iOS/Classes/GridSpot.swift
