+++ Sources/iOS/Classes/GridSpotCell.swift
