+++ Sources/iOS/Classes/ListSpot.swift
