+++ Sources/iOS/Classes/ListSpotCell.swift
