+++ Sources/iOS/Extensions/ListAdapter+Extensions+iOS.swift
