+++ Sources/iOS/Extensions/SpotsController+UIScrollViewDelegate.swift
